+++ client/components/Sidebar.tsx
@@ -1,9 +1,11 @@
 import Communities from "./Communities"
+import { Button } from "@mui/material";
 
 const Sidebar = () => {
   return (
     <>
       <Communities />
+      <Button href="/communities/create">Create</Button>
     </>
   )
 }
